refactor repository decoration providers
